moved validation out of router into separate file
