+++ server/middleware/validator.js
@@ -1,0 +1,21 @@
+module.exports.validRegistration = () => {
+  return function(req, res, next) {
+    const username = req.body.username || ''
+    const password = req.body.password || ''
+    const password2 = req.body.password2 || ''
+    const errorsArray = []
+    let isValid = true
+    if (password !== password2) {
+      errorsArray.push('Passwords must match')
+      isValid = false
+    } else if (password === "") {
+      errorsArray.push('Password cannot be empty')
+      isValid = false
+    }
+    // TO DO more validation??
+    if (!isValid) {
+      return res.json({ error: true, errorsArray })
+    }
+    next()
+  }
+}
